requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
black==23.11.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
fastapi==0.104.1
flake8==6.1.0
h11==0.14.0
idna==3.4
mccabe==0.7.0
mypy==1.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pycodestyle==2.11.1
pydantic==2.5.1
pydantic_core==2.14.3
pyflakes==3.1.0
PyJWT==2.8.0
pymongo==4.6.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
